Add timer to refresh tokens
The default token timeout is 3600 seconds (1 hour), so if the download
takes too much time, the token times out and all remaining files will be
a 59-byte error message as documented in issue #35.  To address this
issue and to make the program more reliable, the previously used log of
completed recordings is removed since it does not save much time and
because it seemed to often erroneously classify recordings as
successfully downloaded even when there were errors, such as the token
timeout error.

Instead, this version uses the file size and checks to see if the size
of the downloaded file matches the file size as reported by the zoom
API.  If the file exists, and the file length matches, this is assumed
to mean that the file has already successfully been downloaded.
Otherwise, the file is downloaded anew.

Signed-off-by: Ed Beroset <beroset@ieee.org>